Move space_time_region to models
