+++ src/virtualship/models/space_time_region.py
